building new parse

+++ parse.sh
@@ -1,12 +1,18 @@
 #!/bin/bash
 set -e
 
+if [ ! -e /usr/bin/parse ]; then
+	sudo cp -u parse.sh /usr/bin/parse
+	echo -e "\\nparse is installed"
+fi
+
 format='.*(webrip|avi|flv|wmv|mov|mp4|mkv|3gp|webm|m4a|m4v|f4a|f4v|m4b|m4r|f4b).*</a>'
 directory='.temp'
 ct=.ct; ct2=.ct2; out=.out; out2=.out2
 arch=$(uname)
 
 cleanup(){
+	if [ -e .logs ]; then echo -e "logs:" && cat -n .logs; rm .logs; fi
 	if [ -e .out ]; then rm .out*; fi
 	if [ -e .ct ]; then rm .ct*; fi
 	if [ -e .ct2 ]; then rm .ct*; fi
@@ -17,8 +23,20 @@ sig_abort(){
 	cleanup && echo -e "\\naborted.\\n" && exit 0
 }
 
+connect(){
+	wget -q --spider google.com && exitStatus=$?||exitStatus=4
+	if [ $exitStatus == 0 ]; then
+		if grep -q "404" .wget;then echo -e "\\n:: error: 404 not found.\\n"
+		elif grep -q "Name or service not known" .wget;then
+			echo -e "\\n:: service unknown: '$url'\\n"
+		fi
+	else
+		echo -e "\\n:: no internet connection.\\n"
+	fi
+	cleanup && exit
+}
 dl(){
-	wget -k "$positional_parameter" -o .wget -O $ct #||connect
+	wget -k "$positional_parameter" -o .wget -O $ct||connect
 	if ! grep -qE "$format" $ct; then echo -e ":: error: $positional_parameter\\n"; cleanup; exit; fi
 	if grep -qiE '(&amp|&darr)' $ct; then
 		grep -vE '.*(amp|darr).*' $ct > $ct2
@@ -34,26 +52,29 @@ title(){
 	if [ $no_title ]; then
 		return
 	else
-		{ if grep -qioE "s[0-9][0-9]" $ct; then
-			_a=$(grep -ioE "s[0-9][0-9]" $ct|head -1|sed -e "s/[S-s]/<h3>Season /" -e 's/$/<\/h3>/')
-			echo -e "$_a"
-		  fi; } >> $out
+		{	if grep -qioE "s[0-9][0-9]" $ct; then
+				_a=$(grep -ioE "s[0-9][0-9]" $ct|head -1|sed -e "s/[S-s]/<h3>Season /" -e 's/$/<\/h3>/')
+				echo -e "$_a"
+			else
+				movies=yes
+		  	fi
+		} >> $out
 	fi
 }
 place_in_order(){
+	if [ $movies ]; then return; fi
 	c=1; d=1
 	lines=$(wc -l<$out)
 	if grep -q "Season" $out; then grep "Season" $out >> $ct2; lines=$((lines-1)); fi
 	while true; do
-		if [ $c != $lines ]; then
+		if [ $c -le $lines ]; then
 			if [ $d -gt 9 ]; then e="(e$c|e0$c|-e$c|-$c)"
 			else e="(e$c|e0$c|e00$c|-e0$c|-0$c)"
 			fi
 			grep -iwE ".*$e" $out >> $ct2||true
 			c=$((c+1)); d=$((d+1))
 		else
-			if [ ! -e $ct2 ]; then echo -e ":: error: cannot place in order."; cleanup && exit; fi
-			mv $ct2 $out
+			if [ -e $ct2 ]; then mv $ct2 $out; else echo "place in order failed: $positional_parameter" >> .logs; fi
 			break
 		fi
 	done
@@ -100,9 +121,14 @@ final(){
 	fi
 }
 sort_(){
-	if ! grep -qE "(480p|720p|1080p)" $ct; then
-		echo >> $out
-		grep -iowE "$format" $ct|sed 's:.*<a:<a:' >> $out
+	if [ "$1" == final ]; then ct=$file; grep -o '.*Season.*' "$ct"|head -1 >> $out; fi
+
+	if ! grep -qE "(480p|720p|1080p)" $ct; then echo >> $out
+		if [ "$1" == final ]; then
+			grep -iowE "$format" $ct >> $out
+		else
+			grep -iowE "$format" $ct|sed 's:.*<a:<a:' >> $out
+		fi
 	else
 		#get 480p, 720p & 1080p only
 		if [ $first_pixel ]; then f=(480p); if ! grep -q "$first_pixel" $ct; then echo -e "\\n:: null: $first_pixel.\\n"; cleanup && exit; fi
@@ -110,43 +136,48 @@ sort_(){
 		elif [ $third_pixel ]; then f=(1080p); if ! grep -q "$third_pixel" $ct; then echo -e "\\n:: null: $third_pixel.\\n"; cleanup && exit; fi
 		else f=(480p 720p 1080p)
 		fi
+		touch $ct2
 		for f in "${f[@]}"; do
-			if grep -qioE "$f.*$format" $ct; then
-				grep -owE ".*$f.*$format" $ct > $ct2
-				if grep -qE "(x264|x265)" $ct2; then
-					perg(){
-						size=$(wc -l<.ct3)
-						if [ "$size" -ge 1 ]; then
-							{
-								cat .ct3
-							} >> $out
-						fi
-						rm .ct3
-						}
-					grep -vE "$f.*(x264|x265)" $ct2 >> .ct3 && perg || true
-				else
-					size=$(wc -l<$ct2)
+			if [ "$1" == final ]; then
+				if grep -qioE "$f" "$file"; then grep -E "$f" "$file" > $ct2; else continue; fi
+			else
+				if grep -qioE "$f.*$format" $ct; then grep -owE ".*$f.*$format" $ct > $ct2; else continue; fi
+			fi
+			if grep -qE "(x264|x265)" $ct2; then
+				flow(){
+					size=$(wc -l<.ct3)
 					if [ "$size" -ge 1 ]; then
-						{
-							grep -owE "($f|$format)" $ct2
+						{	if [ "$1" == final ]; then echo -e "\\n$f\\n"; fi
+							cat .ct3
 						} >> $out
 					fi
-				fi
-				#get x264 only
-				if grep -qioE "$f.*x264" $ct2; then
-					{
-						grep -oiE ".*$f.*x264" $ct2
+					rm .ct3
+				}
+				grep -vE "$f.*(x264|x265)" $ct2 >> .ct3 && flow || true
+			else
+				{	if [ "$1" == final ]; then echo -e "\\n$f\\n"; grep -E "$f" $ct2
+					else
+						size=$(wc -l<$ct2)
+						if [ "$size" -ge 1 ]; then grep -owE "($f|$format)" $ct2; fi
+					fi
 					} >> $out
-				fi
-				#get x265 only
-				if grep -qioE "$f.*x265" $ct2; then
-					{
-						grep -oiE ".*$f.*x265" $ct2
-					} >> $out
-				fi
+					continue
+			fi
+			#get x264 only
+			if grep -qioE "$f.*x264" $ct2; then
+				{	if [ "$1" == final ]; then echo -e "\\n$f x264\\n"; fi
+					grep -oiE ".*$f.*x264" $ct2
+				} >> $out
+			fi
+			#get x265 only
+			if grep -qioE "$f.*x265" $ct2; then
+				{	if [ "$1" == final ]; then echo -e "\\n$f x265\\n"; fi
+					grep -oiE ".*$f.*x265" $ct2
+				} >> $out
 			fi
 		done
 	fi
+	if [ "$1" == final ]; then return; fi
 	if [ -e $ct2 ]; then rm .ct[2-3];fi
 	place_in_order
 }
@@ -185,7 +216,7 @@ parsing(){
 			cleanup && exit
 		else
 			header top
-			for file in "$directory"/*;do final;done
+			for file in "$directory"/*;do sort_ final;done
 			header
 			#if [ "$arch" == Darwin ];then
 			#	pbcopy < $out;else
